Added ability to override saving/retrieving data by having ArrayAccess and magic methods check for get/set methods.
For instance, if you have $entity['some_variable'], the object will check for the $entity:->getSomeVariable() method. If it exists, it will use this instead. The same thing occurs for setting a variable: $entity->someVar = ''foo" would cause the object to check for $entity->setSomeVar() and if it exists would call $entity->setSomeVar($value) instead of setting the variable directly.